Fix #2: Eclipse plugins can not be updated
Updates will be installed during container build. Since the p2 director has no
update function, the plugins of the eclipse package must be uninstalled before
reinstallation. Therefore, a new section 'uninstall' was added to
'plugins.yml'.

Manual update procedures using the "Check for Updates' button is not
recommended, because the changes are not persistent.